bbdd added,crear,modificar,borrar, HU1...HU4 DONE

+++ src/main/java/com/example/todolist/model/repositories/ItemListRepository.java
@@ -3,5 +3,5 @@ package com.example.todolist.model.repositories;
 import com.example.todolist.model.entities.ItemList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemListRepository extends JpaRepository<ItemList, Long> {
+public interface ItemListRepository extends JpaRepository<ItemList, Integer> {
 }
